Simplify abstract query functionality
No need to have the trait when all queries have a base class.
